refactor: mv test/ to test-util/
