+++ src/emitter/index.test.ts
@@ -1,6 +1,6 @@
 import { readFile, stat } from 'fs/promises';
 import { jest } from '@jest/globals';
-import { createFixtures, exists, fakeToken, getFixturePath, waitForAsyncTask } from '../test/util.js';
+import { createFixtures, exists, fakeToken, getFixturePath, waitForAsyncTask } from '../test-util/util.js';
 import { emitGeneratedFiles, getRelativePath, isSubDirectoryFile } from './index.js';
 
 // eslint-disable-next-line @typescript-eslint/no-empty-function
