+++ src/runner.test.ts
@@ -8,7 +8,7 @@ import { jest } from '@jest/globals';
 import chalk from 'chalk';
 import dedent from 'dedent';
 import type { Watcher } from './runner.js';
-import { createFixtures, exists, getFixturePath, waitForAsyncTask } from './test/util.js';
+import { createFixtures, exists, getFixturePath, waitForAsyncTask } from './test-util/util.js';
 
 const require = createRequire(import.meta.url);
