+++ src/util.test.ts
@@ -1,6 +1,6 @@
 import { join } from 'path';
 import { hasProp, isObject, isSystemError, unique, uniqueBy } from '../src/util.js';
-import { createFixtures, getFixturePath } from './test/util.js';
+import { createFixtures, getFixturePath } from './test-util/util.js';
 import { isMatchByGlob, exists } from './util.js';
 
 function fakeSystemError({ code }: { code: string }) {
